requirements.txt
# === Core Framework ===
fastapi==0.115.5
uvicorn[standard]==0.32.0

# === Validation & Settings ===
pydantic==2.8.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
structlog==24.4.0
orjson==3.10.7
httpx==0.28.1

# === Data / Export ===
numpy==1.26.4
pandas==2.2.3
openpyxl==3.1.5

# === HTML cleaning ===
beautifulsoup4==4.12.3
lxml==5.3.0

# === RAG / Embedding ===

sentence-transformers==2.6.1

# === LangChain Ecosystem ===
langchain==0.3.11
langchain-core==0.3.27
langchain-community==0.3.11
langchain-openai==0.2.3
pinecone==5.0.1
langgraph>=0.2.33,<0.3.0
pinecone==5.0.1
langchain-pinecone 
langchain-aws==0.2.10




# === Web search / Tavily ===
tavily-python==0.7.14
